+++ lib/Elastica/Search.php
@@ -477,12 +477,14 @@ class Search
 
         $query = $this->getQuery();
         $path = $this->getPath();
+        $params = $this->getOptions();
+        $params[self::OPTION_SEARCH_TYPE] = self::OPTION_SEARCH_TYPE_COUNT;
 
         $response = $this->getClient()->request(
             $path,
             Request::GET,
             $query->toArray(),
-            array(self::OPTION_SEARCH_TYPE => self::OPTION_SEARCH_TYPE_COUNT)
+            $params
         );
         $resultSet = ResultSet::create($response, $query);